zmiany w nawigation
