+++ src/App.js
@@ -3,14 +3,14 @@ import './App.css';
 import {Grid, Row, Col} from 'react-bootstrap';
 import Menu from './Components/ElementsPage/Menu.js';
 import Sliders from './Components/ElementsPage/Sliders.js';
-import Detalis from './Components/ElementsPage/Detalis.js';
+import Detalis from './Components/ElementsPage/Main.js';
 import Footer from './Components/ElementsPage/Footer.js';
 
     class App extends Component {
 
         render() {
             return (
-                <Grid>
+                <Grid fluid>
                     <Row className="show-grid">
                         <Col className="App_nav" xs={12} md={12}>
                             <Menu />
